creates main component
